v1 probe on the frozen-candidate wire spec (oep-spec 567f57b), before hardware tests: numbers from the vendored registry (src/OepV1Registry.h), one request-tail helper (critical -> unsupported with the tag as received, non-critical -> ignored 0x7F, 0x7F / 0xFF malformed), confirm with a revision range (window u32), list / describe u16 with oep.core listed, reject no_connection / unsupported, force ends subscriptions, subscribe takes all three fields, heartbeat 0 -> 1000 ms; wires and riscv-dm / arm-adi at revision 1 (status enum, failures as completed failed/partial, attach speed_hz / existing / max_speed, RVSWD attach writes only at the slowest period until the speed is chosen by reads, dmi n + poll values + done rule, halt idempotent, resume re-issue rules, run with u32 timeout and n_out); console rev 1 (existing stream, marks by serial, closed streams readable), gpio / uart rev 1 on a shared PositionStream, capture rev 1 (repeat reads found by serial, wrap-safe); examples moved to the v1 fixtures (the V003 probe has no capture now)

+++ src/OepFrame.h
@@ -10,7 +10,7 @@ namespace oep {
 // What an endpoint accepts per frame and keeps in flight (core confirm reports it).
 struct Limits {
   uint16_t max_frame;
-  uint16_t window_bytes;
+  uint32_t window_bytes;   // v1 confirm reports a u32; v0 a u16
   uint8_t max_inflight;
 };
